backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-multipart
shap
matplotlib
asttokens==3.0.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
fonttools==4.61.1
imbalanced-learn==0.14.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
llvmlite==0.46.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numba==0.63.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==65.5.0
shap==0.50.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.2.14
